+++ server/src/main/kotlin/suwayomi/tachidesk/server/util/Chromium.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.network.interceptor
+package suwayomi.tachidesk.server.util
 
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.json.Json
